fixed most pr issues
